+++ src/test/java/stringcalculator/StringAddCalculatorTest.java
@@ -2,7 +2,7 @@ package stringcalculator;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static stringcalculator.ErrorMessageConstants.NON_POSITIVE_INTEGER_INPUT_ERROR;
+import static stringcalculator.constants.ErrorMessageConstants.NON_POSITIVE_INTEGER_INPUT_ERROR;
 
 import org.junit.jupiter.api.Test;
